+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 172:
 * Tidy up websocket stream javadocs
 * Fix move-only arguments in bind_handler
 * Fix http::parser constructor javadoc
+* Tidy up test::stream javadocs
 
 --------------------------------------------------------------------------------